// src/interfaces/model/account.interface.ts
export interface IAccountCreate {
	id?: number;
	clientId?: number;
	timesheetId?: number;
	n: string;
	position: string;
	type: string;
	affectation: string;
	serviceMonth: string;
	monthlySalaryWithHousingAndTravel: number;
	daysWorked: number;
	dailyCost: number;
	shouldBeInvoiced: number;
	invoiced?: number;
	toBeInvoicedBack?: number;
	poNumber?: string;
	poDate?: Date;
	invoiceNumber?: string;
	invoiceLodgingDate?: Date;
	invoiceAmount?: number;
	salaryPaid?: number;
	bonus1Name?: string;
	bonus1?: number;
	poBonus1?: number;
	invoiceNumberPOBonus1?: string;
	bonus2Name?: string;
	bonus2?: number;
	poBonus2?: number;
	invoiceNumberPOBonus2?: string;
	bonus3Name?: string;
	bonus3?: number;
	poBonus3?: number;
	invoiceNumberPOBonus3?: string;
	additionalAmount?: number;
	additionalPOBonus?: string | null;
	additionalInvoiceNumberPO?: string | null;
	dateSalaryPaid?: Date;
	comments?: string;
	isGeneratedInvoice: boolean;
	additionalBonusNames?: string;
	poNumberBonus1?: string;
	poNumberBonus2?: string;
	poNumberBonus3?: string;
}

export interface AccountUpdate {
	employeeId: string;
	invoiced?: number | null;
	invoiceNumber?: string | null;
	invoiceAmount?: number | null;
	salaryPaid?: number | null;
	invoiceLodgingDate?: Date | null;
	dateSalaryPaid?: Date;
	comments?: string | null;
	isGeneratedInvoice: boolean;
	poNumber?: string;
	poDate?: Date;
	bonus1Name?: string;
	bonus1?: number;
	poBonus1?: number;
	invoiceNumberPOBonus1?: string;
	bonus2Name?: string;
	bonus2?: number;
	poBonus2?: number;
	invoiceNumberPOBonus2?: string;
	bonus3Name?: string;
	bonus3?: number;
	poBonus3?: number;
	invoiceNumberPOBonus3?: string;
	additionalAmount?: number;
	additionalPOBonus?: string | null;
	additionalInvoiceNumberPO?: string | null;
	additionalBonusNames?: string;
	poNumberBonus1?: string;
	poNumberBonus2?: string;
	poNumberBonus3?: string;
}
export interface AccountAttributes {
	id?: number;
	clientId?: number;
	timesheetId?: number;
	employeeId?: number;
	n: string;
	isGeneratedInvoice?: boolean;
	position: string;
	type: string;
	affectation: string;
	serviceMonth: string;
	monthlySalaryWithHousingAndTravel: number;
	daysWorked: number;
	dailyCost: number;
	shouldBeInvoiced: number;
	invoiced?: number;
	toBeInvoicedBack?: number;
	poNumber?: string;
	poDate?: Date;
	invoiceNumber?: string;
	invoiceLodgingDate?: Date;
	invoiceAmount?: number;
	salaryPaid?: number;
	bonus1Name?: string;
	bonus1?: number;
	poBonus1?: number;
	invoiceNumberPOBonus1?: string;
	bonus2Name?: string;
	bonus2?: number;
	poBonus2?: number;
	invoiceNumberPOBonus2?: string;
	bonus3Name?: string;
	bonus3?: number;
	poBonus3?: number;
	invoiceNumberPOBonus3?: string;
	additionalAmount?: number;
	additionalPOBonus?: string | null;
	additionalInvoiceNumberPO?: string | null;
	additionalBonusNames?: string;
	poNumberBonus1?: string;
	poNumberBonus2?: string;
	poNumberBonus3?: string;
	dateSalaryPaid?: Date;
	comments?: string;
	createdAt?: Date | string;
	createdBy?: number;
	updatedAt?: Date | string;
	updatedBy?: number;
	deletedAt?: Date | string;
	createdatutc?:string;
	updatedatutc?:string;
	deletedatutc?:string;
}